#13 [FEAT] 지도 인구 혼잡도 api 연동

+++ data/src/main/java/com/gdg/data/mapper/ResponseCongestionDtoMapper.kt
@@ -4,7 +4,7 @@ import com.gdg.data.dto.response.CongestionResponseDto
 import com.gdg.domain.entity.CongestionEntity
 
 fun CongestionResponseDto.toCongestionEntity() = CongestionEntity(
-    id = areaId,
+    id = areaId.toLong(),
     name = areaNm,
     level = areaCongestLvl,
     message = areaCongestMsg,
